+++ packages/ui/src/PoweredByLinen/index.module.scss
@@ -1,4 +1,4 @@
-@import '../../../colors.scss';
+@import '../colors.scss';
 
 .link {
   align-items: center;
